feat: enable metricsgen to be run as go sdk
